+++ custom_components/hph/binary_sensor.py
@@ -61,8 +61,6 @@ class HphTemplateBinarySensor(BinarySensorEntity):
         self._attr_unique_id = unique_id
         self.entity_id = f"binary_sensor.{unique_id}"
         self._attr_translation_key = unique_id
-        raw = definition.get("name", unique_id)
-        self._attr_name = raw.removeprefix("HeatPump Hero ").removeprefix("HPH ")
         self._attr_icon = definition.get("icon")
 
         if "device_class" in definition: