+++ custom_components/hph/button.py
@@ -87,7 +87,6 @@ class HphFacadeButton(FacadeProxyMixin, ButtonEntity):
         self._attr_unique_id = unique_id
         self._attr_translation_key = unique_id
         self.entity_id = f"button.{unique_id}"
-        self._attr_name = cfg.get("name", unique_id)
         self._attr_icon = cfg.get("icon")
         self._writer_id = f"text.{cfg['writer']}"
         self._attr_available = False